% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alien.test}
\alias{alien.test}
\title{Alien test dataset}
\format{A data.table with 8 rows and 5 variables:
\describe{
  \item{SkinColor}{Skin color of the individual}
  \item{IQScore}{IQ score of the individual}
  \item{Cat1}{Categorical descriptor}
  \item{Cat2}{Categorical descriptor}
  \item{Cat3}{Categorical descriptor}
}}
\usage{
alien.test
}
\description{
A dataset describing features of living beings
}
\keyword{datasets}

