% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alien.train}
\alias{alien.train}
\title{Alien training dataset}
\format{A data.table with 8 rows and 6 variables:
\describe{
  \item{SkinColor}{Skin color of the individual}
  \item{IQScore}{IQ score of the individual}
  \item{Cat1}{Categorical descriptor}
  \item{Cat2}{Categorical descriptor}
  \item{Cat3}{Categorical descriptor}
  \item{IsAlien}{Is this being an alien?}
}}
\usage{
alien.train
}
\description{
A dataset describing features of living beings and whether or not they are an alien
}
\keyword{datasets}

