% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_roc.R
\name{auc_roc}
\alias{auc_roc}
\title{Area Under the ROC Curve}
\usage{
auc_roc(preds, actuals, returnDT = FALSE)
}
\description{
Area Under the ROC Curve
}
\examples{
preds <- c(.1, .3, .3, .9)
actuals <- c(0, 0, 1, 1)
auc_roc(preds, actuals)
auc_roc(preds, actuals, returnDT=TRUE)
}

