% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_calc_auc_pr.R
\name{bin_calc_auc_pr}
\alias{bin_calc_auc_pr}
\title{Calculate AUC-PR}
\usage{
bin_calc_auc_pr(actual, predicted)
}
\arguments{
\item{actual}{original data values - binary.}

\item{predicted}{predicted values in probabilities or binary.}
}
\value{
Returns the area under the curve
}
\description{
Calculates Area under the precision-recall curve
}
