% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_conf_matrix.R
\name{bin_conf_matrix}
\alias{bin_conf_matrix}
\title{Confusion Matrix for a binary model}
\usage{
bin_conf_matrix(actual, predicted, threshold = 0.5)
}
\arguments{
\item{actual}{original data values - binary.}

\item{predicted}{predicted values in probabilities or binary.}

\item{threshold}{threshold at which predicted probabilities are to classified.}
}
\value{
Returns the confusion matrix as a dataframe
}
\description{
Computes confusion matrix along with other evaluation parameters for binary
classification problem
}
