% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_model_eval.R
\name{bin_model_eval}
\alias{bin_model_eval}
\title{Binary Model Evaluation}
\usage{
bin_model_eval(actual, predicted)
}
\arguments{
\item{actual}{original data values - binary.}

\item{predicted}{predicted values in probabilities or binary.}
}
\value{
a list of AUC-ROC, AUC-PR, ROC and PR charts, threshold tuned df
}
\description{
Evaluates model performance parameters - AUC-ROC, AUC-PR, ROC and PR charts, threshold tuned df
}
