% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_thres_eval.R
\name{bin_thres_eval}
\alias{bin_thres_eval}
\title{Threshold Evaluation}
\usage{
bin_thres_eval(actual, predicted, threshold = 0.5)
}
\arguments{
\item{actual}{original data values - binary.}

\item{predicted}{predicted values in probabilities or binary.}

\item{threshold}{threshold at which predicted probabilities are to classified.}
}
\value{
Returns True/False Positive/Negatives
}
\description{
Evaluate the model for a given threshold
}
