% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallFunctions.R
\name{CheckCateg}
\alias{CheckCateg}
\title{check if a variable is categorical variable}
\usage{
CheckCateg(x, maxLevel = 10)
}
\arguments{
\item{x}{a vector of data}

\item{maxLevel}{A number to specify the max number of categories for categorical data.}
}
\value{

}
\description{
check if a variable is categorical variable
}
\examples{
CheckCateg(x)

}

