% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complexityCurve.R
\name{complexityCurve}
\alias{complexityCurve}
\title{Generate complexity curve data}
\usage{
complexityCurve(task, learner, ps)
}
\arguments{
\item{task}{A mlr task}

\item{learner}{A mlr learner}

\item{ps}{A parameter set over which to optimize model}
}
\description{
Generate complexity curve data
}
\author{
Adam Acosta
}

