% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complexityCurve.R
\name{complexityCurvePlot}
\alias{complexityCurvePlot}
\title{Plot complexity curve}
\usage{
complexityCurvePlot(data, x, ggvis = FALSE)
}
\arguments{
\item{data}{A data.frame with training and testing error to plot}

\item{x}{Name of the complexity parameter to plot on the X axis}

\item{ggvis}{logical Whether or not to use ggvis; default FALSE}
}
\value{
A ggplot object if ggvis = TRUE, otherwise no return value
}
\description{
Plot complexity curve
}
\author{
Adam Acosta
}

