% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_matrix.R
\name{conf_matrix}
\alias{conf_matrix}
\title{Binary Model Evaluation}
\usage{
conf_matrix(actual, predicted, threshold = 0.5, fn_weight = 10,
  fp_weight = 1)
}
\arguments{
\item{actual}{original data values - binary.}

\item{predicted}{predicted values in probabilities or binary.}

\item{threshold}{threshold at which predicted probabilities are to classified.}

\item{fn_weight}{cost of false negatives when calculating asymmetric weight. Defaulted to 10}

\item{fp_weight}{cost of false positives when calculating asymmetric weight. Defaulted at 1.}
}
\value{
Returns a list that contains the confusion matrix as a dataframe, and another dataframe
that contains evaluation parameters. Visit \href{https://en.wikipedia.org/wiki/Sensitivity_and_specificity}{Wikipedia}  for definitions
}
\description{
Computes confusion matrix along with other evaluation parameters for binary
classification problem
}
