% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExploreFunc.R
\name{DistriSum}
\alias{DistriSum}
\title{Generate distribution plot and summary}
\usage{
DistriSum(x, outDir = NULL, colCalss = FALSE, maxLevel = 10, ...)
}
\arguments{
\item{x}{Input data ML.data}

\item{outDir}{Path to directory to save the plot and table.
If it is specified, the plots will be saved. Default is NULL}

\item{maxLevel}{A number to specify the max number of categories
for categorical data.}

\item{...}{Parameters in getData}

\item{colClass}{A boolen. If it is TRUE, the ditribution will colored by
Classes. Default is FALSE.}
}
\value{
plots and summary for continuous variable and catergorical variables.
}
\description{
Generate distribution plot and summary
}
\examples{
DistriSum(x)

}

