% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_nn.R
\name{gen_nn}
\alias{gen_nn}
\title{Generate Artifical Neural Network}
\usage{
gen_nn(data.set, ..., cName = "res", logs = FALSE, fold.info = c(10, 7))
}
\arguments{
\item{data.set}{A data set that contains scaled data
and a vector of results as a column in the data frame.}
}
\value{
A neural network model built from \code{totalData}
 which is of class 'nn'.
}
\description{
A function that takes a scaled data set
 builds a neural network and reports on the accuracy
 of the built model.
}
