% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAccess.R
\name{GenData}
\alias{GenData}
\title{GenData}
\usage{
GenData(data = NULL, class.var = NULL, pred.var = NULL, ...)
}
\arguments{
\item{data}{A data frame or path of file.}

\item{class.var}{A vecter to specify the variable in data is classification information.}

\item{pred.var}{A vecter to specify the variable in data is prediction variable.}

\item{...}{Parameters for read file}
}
\value{
ML.data class
}
\description{
GenData
}

