% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folds.R
\name{get_folds}
\alias{get_folds}
\title{Get Folds}
\usage{
get_folds(target, folds = 5L, stratified = FALSE, seed = NULL)
}
\arguments{
\item{target}{A vector of target values}

\item{folds}{How many folds}

\item{stratified}{Should target classes be split proportionally (how to handle numeric target?)}

\item{seed}{Random number seed}
}
\description{
Map a target vector to a set of FoldIDs of equal length (or nearly equal length)
}
\details{
Convenient method for mapping each instance to a FoldID, potentially with stratification
}
\examples{
get_folds(iris$Species, 5L)
get_folds(iris$Species, 5L, stratified=TRUE)

}

