% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/methodAccess.R
\docType{methods}
\name{getClass}
\alias{getClass}
\alias{getClass,ML.data-method}
\title{getClass}
\usage{
getClass(object, IDs)

\S4method{getClass}{ML.data}(object, IDs)
}
\arguments{
\item{object}{ML.data object}

\item{IDs}{A vector of IDs for each data object. When specified, class info only
with these IDs will be output}
}
\value{
Classification info.
}
\description{
getClass
}
\examples{

}

