% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/methodAccess.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,ML.data-method}
\title{getData}
\usage{
getData(object, IDs, Cols, Filters)

\S4method{getData}{ML.data}(object, IDs, Cols, Filters)
}
\arguments{
\item{object}{ML.data object}

\item{IDs}{A vector of IDs for each data object. When specified, data only
with these IDs will be output}

\item{Cols}{A vector of colnames. When specified, data only for
these colunm will be output.}

\item{Filters}{A string like "Col1 == 1 | Col2 == 2& Col3 >2". This will
filter data based on the criteria.}
}
\value{
data for machine learning.
}
\description{
getData
}
\examples{

}

