% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_impurity.R
\name{gini_impurity}
\alias{gini_impurity}
\title{Gini Impurity}
\usage{
gini_impurity(vals)
}
\arguments{
\item{vals}{A vector of values. Values can be given as raw instances like c("red", "red", "blue", "green") or as a named vector
of class frequencies like c(red=2, blue=1, green=1)}
}
\description{
Calculates the Gini Impurity of a set. Gini Impurity is a measure of how often a randomly chosen element from the set would be
incorrectly labeled if it was randomly labeled according to the distribution of labels in the set. Values may be provided as a
vector or via a pair of parameters - a data.table object and the name of the column to bin.
}
\details{
Returns a real value in the range [0, 1]
}
\examples{
gini_impurity(c("red", "red", "blue", "green"))
gini_impurity(c(red=2, blue=1, green=1))

}

