% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/GrDataKNN.R
\name{GrDataKNN}
\alias{GrDataKNN}
\title{GrDataKNN}
\usage{
GrDataKNN(X, Y, k)
}
\arguments{
\item{X}{A data frame or a matrix where rows are observations and columns are features.}

\item{Y}{A vector with labels for each row in \code{data}.}

\item{k}{Number of neighbors that the classifier should use. It has to be an odd number.}
}
\value{
A List with 3 data sets. Boundries are the points from the region where there is a change of categories, Popoints are all the points in the graph and the category and truedata the original points.
}
\description{
Finds the boundries of each category through the KNN method.
}
\examples{
# create artificial dataset
x1 <- c(978783,1052488,1039495,1056795,1125545,1536011,1616461,1388507,1608121,1416574,1704919,2653310,2208399,1896304,1957401)
x2 <- c(12645844,12343453,4137266,12229065, 12554668,8856611,12137668,11545424,9253718,8863474,15145969,12793921,14593861,15161586,15785243)
Y <- c(rep("Cats",5),rep("Dogs",5),rep("Lions",5))
X <- as.data.frame(cbind(x1,x2))
data <- cbind(X,Y)
# get the kNN predictions for the test set
GrDataKNN(X=data[,1:2],Y=data[,3],k=5)
 ggplot()+
 geom_point(aes(x=BouPoints[,1],y=BouPoints[,2]),data=BouPoints, lwd=1)+
 geom_point(aes(x=Popoints[,1],y=Popoints[,2], colour=category),data=Popoints, alpha=1/4, pch=1, lwd=1)+
 geom_point(aes(x=TrueData[,1], y=TrueData[,2], colour=category), data=TrueData, pch=18, lwd=3)+
 xlab("Weight") +
 ylab("Height")
}

