% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learningCurve.R
\name{learningCurve}
\alias{learningCurve}
\title{Generate data for a learning curve plot}
\usage{
learningCurve(data, learner)
}
\arguments{
\item{data}{data.frame Contains training examples}

\item{learner}{A learner object}
}
\value{
A data frame with columns
  \item{Examples}{integer Sample size used}
  \item{Type}{factor with levels 'training' and 'testing'}
  \item{Error}{numeric Mean misclassification error across repeated CV}
}
\description{
Generate data for a learning curve plot
}
\author{
Adam Acosta
}

