% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learningCurve.R
\name{learningCurvePlot}
\alias{learningCurvePlot}
\title{Generate a learning curve plot}
\usage{
learningCurvePlot(data, ggvis = FALSE)
}
\arguments{
\item{data}{A data.frame containing learning curve data}

\item{ggvis}{Logical Should ggvis be used; default FALSE}
}
\value{
A ggplot object if ggvis = TRUE, otherwise no return value
}
\description{
Generate a learning curve plot
}
\author{
Adam Acosta
}

