% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{ML.data-class}
\alias{ML.data-class}
\title{An S4 class contains 5 slots: methylation, expression, sample information, probe
information and gene information. MEE.data class are the main input for main functions.}
\description{
An S4 class contains 5 slots: methylation, expression, sample information, probe
information and gene information. MEE.data class are the main input for main functions.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame contains data for ML.
each column is one individual features}

\item{\code{classes}}{A data.frame contains data for ML. Columns are IDs, CLASS, ...}

\item{\code{pred}}{A data.frame contains prediction. Columns are IDs, PRED, PROB, ...}
}}

