% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msle.R
\name{msle}
\alias{msle}
\title{Mean Square Logarithmic Error}
\usage{
msle(preds = NULL, actuals = NULL, na.rm = FALSE, weights = 1)
}
\arguments{
\item{preds}{A vector of prediction values in [0, 1]}

\item{actuals}{A vector of actuals values in {0, 1}, or {FALSE, TRUE}}

\item{na.rm}{Should (prediction, actual) pairs with at least one NA value be ignored?}

\item{weights}{Optional vectors of weights}
}
\description{
Calculate Mean-Square-Logarithmic Error (Deviation)
}
\details{
Calculate Mean-Square-Logarithmic Error (Deviation)
}
\examples{
preds <- c(1.0, 2.0, 9.5)
actuals <- c(0.9, 2.1, 10.0)
msle(preds, actuals)

}
