% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{plotVI.pipe_result.keras}
\alias{plotVI.pipe_result.keras}
\title{plotVI.pipe_result.keras}
\usage{
plotVI.pipe_result.keras(
  res,
  vi = c("ratio", "diff", "raw"),
  dispersion = c("sd", "se", "ci")
)
}
\arguments{
\item{dispersion}{sd, se or ci. Metric of the dispersion bars}
}
\description{
plotVI.pipe_result.keras
}
