% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{plotVI.pipe_result}
\alias{plotVI.pipe_result}
\title{plotVI.pipe_result}
\usage{
plotVI.pipe_result(
  res,
  vi = c("ratio", "diff", "raw"),
  dispersion = c("sd", "se", "ci")
)
}
\arguments{
\item{res}{a \code{pipe_result} object.}

\item{vi}{output type . For ratio and diff, performance is compared with the full model without any permutation.}

\item{dispersion}{sd, se or ci. Metric of the dispersion bars}
}
\description{
plotVI.pipe_result
}
