% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest_helpers.R
\name{predict_randomForest}
\alias{predict_randomForest}
\title{Title}
\usage{
predict_randomForest(
  modelRF,
  predInput,
  scaleInput = FALSE,
  col_means_train,
  col_stddevs_train,
  filename = "",
  tempdirRaster = NULL,
  nCoresRaster = 2
)
}
\arguments{
\item{predInput}{data.frame or raster with colnames or layer names matching the expected input for modelRF}

\item{tempdirRaster}{}
}
\description{
Title
}
