% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_position.R
\name{relative_position}
\alias{relative_position}
\title{Relative Position}
\usage{
relative_position(vals)
}
\arguments{
\item{vals}{vector of values}
}
\description{
Scale a vector of values to the range [0, 1] based on rank/position
}
\details{
Values are ranked and then scaled to the range [0, 1]. Ties result in the same relative position 
(e.g. \code{relative_position(c(1, 2, 2, 3))} returns the vector \code{c(0.0 0.5 0.5 1.0))}. NAs remain as NAs.
}
\examples{
relative_position(1:10)
relative_position(c(1, 2, 2, 3))
relative_position(c(1, NA, 3, 4))

}
