% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{scale_data}
\alias{scale_data}
\title{Scale Data}
\usage{
scale_data(total.results, cLabel = "res")
}
\arguments{
\item{total.results}{A data frame with multiple columns, each one
containing a feature, and a final row with labelled results
i.e. 'W / D / L'.}

\item{cLabel}{A character vector that defines what row the classes 
belong to, usually called 'res'.}
}
\value{
A list object with the following names :
\itemize{
  \item{data : The scaled form of total.results}
  \item{scaler : Another list object containing the scales used to create `data`}
 }
}
\description{
We subtract 1 from the data frame as the data set
 MUST always have a trailing column that contains labelled
 match data with 'W / D / L' etc. Also, one feature is not
 permitted, so total.results must therefore have 3 or more columns.
}
