% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(dt)
}
\arguments{
\item{dt}{A data.table}
}
\description{
Calculates the skewness of each field in a data.table
}
\details{
Counts the frequency of each value in each column, then displays the results in descending order
}
\examples{
library(data.table)
skewness(alien.train)
}

