% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{summary.pipe_result.keras}
\alias{summary.pipe_result.keras}
\title{summary.pipe_result.keras}
\usage{
\method{summary}{pipe_result.keras}(object, ...)
}
\arguments{
\item{object}{a pipe_result.keras object}

\item{...}{parameters to summarize_pred.Raster function.}
}
\description{
summary.pipe_result.keras
}
