% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{summary.pipe_result}
\alias{summary.pipe_result}
\title{summary.pipe_result}
\usage{
\method{summary}{pipe_result}(object, ...)
}
\arguments{
\item{object}{a \code{pipe_result} object.}

\item{...}{parameters to summarize_pred.Raster function.}
}
\description{
summary.pipe_result
}
