% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thres_tune.R
\name{thres_tune}
\alias{thres_tune}
\title{Threshold Tuning}
\usage{
thres_tune(actual, predicted, prob_start = 0, prob_end = 0.5,
  prob_precsn = 0.01, fn_weight = 10, fp_weight = 1)
}
\arguments{
\item{actual}{original data values - binary.}

\item{predicted}{predicted values in probabilities or binary.}

\item{prob_start}{to indicate the value at which the tuning must start. Defaulted to 0}

\item{prob_end}{to indicate at what value of probability the tuning must stop. Defaulted to 0.5}

\item{prob_precsn}{to indicate at what precision each threshold must be incremented. Defaulted to 0.01}

\item{fn_weight}{cost of false negatives when calculating asymmetric weight. Defaulted to 10}

\item{fp_weight}{cost of false positives when calculating asymmetric weight. Defaulted at 1.}
}
\value{
Returns a datafram with model evaluation at each value of threshold
}
\description{
Tunes through a given range of values to determine model performance
}
