% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{transformTS}
\alias{transformTS}
\alias{longToWide.ts}
\alias{wideToLong.ts}
\alias{wideTo3Darray.ts}
\alias{longTo3Darray.ts}
\title{Time series format transformations}
\usage{
longToWide.ts(d, timevar, idCols = NULL)

wideToLong.ts(d, timevar, vars, idCols = NULL, regex_time = ".+")

wideTo3Darray.ts(d, vars, idCols = NULL)

longTo3Darray.ts(d, timevar, idCols = NULL)
}
\arguments{
\item{d}{a data.frame or matrix.}

\item{timevar}{column name for the time.}

\item{idCols}{ids or no time varying columns. Works with column names and indexes.}

\item{vars}{time varying variables in the long format (ie. without the _time part in wide format).}

\item{regex_time}{a regular expression for the time part of the variable names for data in wide format.}
}
\value{
the same data in a wide [var_time columns], long [time + vars columns] or a 3D array [samples, time, vars] format.
}
\description{
For data in wide format, the name of time varying columns follow the pattern \code{var_time}.
}
