% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\name{variableImportance}
\alias{variableImportance}
\title{Variable importance by permutations on predictors}
\usage{
variableImportance(
  model,
  data,
  y,
  repVi = 5,
  variable_groups = NULL,
  perm_dim = NULL,
  comb_dims = FALSE,
  batch_size = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Variable importance by permutations on predictors
}
