% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alien.test}
\alias{alien.test}
\title{Alien test dataset}
\format{A data.table with 8 rows and 5 variables:
\describe{
  \item{SkinColor}{Skin color of the individual}
  \item{IQScore}{IQ score of the individual}
  \item{Cat1}{Categorical descriptor}
  \item{Cat2}{Categorical descriptor}
  \item{Cat3}{Categorical descriptor}
}}
\usage{
alien.test
}
\description{
A dataset describing features of living beings
}
\details{
library(data.table)

alien.test <- data.table::data.table(
  SkinColor=c("white", "green", "brown", "white", "red"),
  IQScore=c(79, 100, 125, 90, 115),
  Cat1=c("type4", "type4", "type3", "type1", "type1"),
  Cat2=c("type5", "type5", "type9", "type8", "type2"),
  Cat3=c("type2", "type2", "type7", "type4", "type4")
)

save(alien.test, file="data/alien_test.rda")
}
\keyword{datasets}

