% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_data.R
\name{bin_data}
\alias{bin_data}
\title{Group data.table into bins}
\usage{
bin_data(dt, col, bins, aggregate = TRUE)
}
\arguments{
\item{dt}{A data.table}

\item{col}{A a column of \code{dt}}

\item{bins}{A vector of binning boundaries}

\item{aggregate}{Should the data be aggregated per bin?}
}
\value{
If \code{aggregate} = TRUE, returns \code{dt} grouped on \code{col} by the specified \code{bins} \cr
taking averages for all other numeric columns. Otherwise returns \code{dt} with bin identifications per row.
}
\description{
Group data.table into bins
}
\examples{
iris.dt <- data.table(iris)
bin_data(dt=iris.dt, col="Sepal.Length", bins=c(4, 5, 6, 7, 8), aggregate=TRUE)

}

