% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot}
\alias{one_hot}
\title{One Hot Encode}
\usage{
one_hot(dt, cols = "auto", dropCols = TRUE, dropUnusedLevels = FALSE)
}
\arguments{
\item{dt}{A data.table}

\item{cols}{Which column(s) should be one-hot-encoded? DEFAULT = "auto" encodes all unordered factor columns}

\item{dropCols}{Should the resulting data.table exclude the original columns which are one-hot-encoded?}

\item{dropUnusedLevels}{Should columns of all 0s be generated for unused factor levels?}
}
\description{
One-Hot-Encode unordered factor columns of a data.table (TODO: make it work on non-factor columns)
}
\details{
One-hot-encoding converts an unordered categorical vector (i.e. a factor) to multiple binarized vectors where each binary vector of
1s and 0s indicates the presence of a class (i.e. level) of the of the original vector.
}
\examples{
library(data.table)

# test dataset
foo <- data.table(
ID=1:5, 
Color=factor(c("green", "red", "red", "blue", "green"), levels=c("blue", "green", "red", "purple")),
Shape=factor(c("square", "triangle", "square", "triangle", "cirlce"))
)

one_hot(foo)
one_hot(foo, dropCols=TRUE)
one_hot(foo, dropCols=FALSE)
one_hot(foo, dropUnusedLevels=TRUE)

}

