% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\name{predict_keras}
\alias{predict_keras}
\title{Title}
\usage{
predict_keras(
  modelNN,
  predInput,
  maskNA = NULL,
  scaleInput = FALSE,
  col_means_train,
  col_stddevs_train,
  batch_size = NULL,
  filename = "",
  tempdirRaster = NULL,
  nCoresRaster = 2
)
}
\arguments{
\item{predInput}{data.frame or raster with colnames or layer names matching the expected input for modelNN}

\item{tempdirRaster}{}
}
\description{
Title
}
