% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_helpers.R
\name{predict_xgboost}
\alias{predict_xgboost}
\title{Title}
\usage{
predict_xgboost(
  modelXGB,
  predInput,
  scaleInput = FALSE,
  col_means_train,
  col_stddevs_train,
  filename = "",
  tempdirRaster = NULL,
  nCoresRaster = 2
)
}
\arguments{
\item{predInput}{data.frame or raster with colnames or layer names matching the expected input for modelXGB}

\item{tempdirRaster}{}
}
\description{
Title
}
